split reference algorithms from  type
